@@ -1,4 +1,4 @@
-package com.exalture.atm
+package com.exalture.atm.landing
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.WindowManager
 import androidx.fragment.app.DialogFragment
+import com.exalture.atm.R
 import kotlinx.android.synthetic.main.dialog_account_type.view.*
 
 class AccountTypeDialog : DialogFragment() {
